refact: SlackHookLogger to bean
