--- piikii-common/src/main/kotlin/com/piikii/common/client/PiikiiRestClient.kt
@@ -1,7 +1,0 @@
-package com.piikii.common.client
-
-import org.springframework.web.client.RestClient
-
-interface PiikiiRestClient {
-    fun build(): RestClient
-}
