+++ piikii-common/src/main/kotlin/com/piikii/common/logutil/HookLogger.kt
@@ -1,5 +1,19 @@
 package com.piikii.common.logutil
 
-interface HookLogger {
-    fun send(message: String)
+import com.piikii.common.logutil.client.HookLoggerConfig
+import org.springframework.context.annotation.Configuration
+import org.springframework.scheduling.annotation.Async
+
+@Configuration
+class SlackHookLogger(
+    private val hookLoggerConfig: HookLoggerConfig,
+) {
+    @Async("hookLoggerTaskExecutor")
+    fun send(message: String) {
+        hookLoggerConfig.getSlackHookLoggerInstance()
+            .post()
+            .body(mapOf("text" to message))
+            .retrieve()
+            .toBodilessEntity()
+    }
 }
