--- piikii-input-http/src/main/kotlin/com/piikii/input/http/client/PiikiiSlackRestClient.kt
@@ -1,16 +1,0 @@
-package com.piikii.input.http.client
-
-import com.piikii.common.client.PiikiiRestClient
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.stereotype.Component
-import org.springframework.web.client.RestClient
-
-@Component
-class PiikiiSlackRestClient(
-    @Value("\${webhook.slack.url}")
-    private val slackHookURL: String,
-) : PiikiiRestClient {
-    override fun build(): RestClient {
-        return RestClient.builder().baseUrl(slackHookURL).build()
-    }
-}
